pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uav_simulator"
version = "0.1.0"
description = "Modular UAV Simulator with high-fidelity physics and control"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
requires-python = ">=3.8"
dependencies = [
    "numpy", "matplotlib", "PyYAML", "scipy", "streamlit", "filterpy"
] 